+++ Runtime/Scalers/AdaptiveFramerate.cs
@@ -6,6 +6,7 @@ namespace UnityEngine.AdaptivePerformance
     public class AdaptiveFramerate : AdaptivePerformanceScaler
     {
         int m_DefaultFPS;
+        int m_FirstTimeStart = 0; // APB-34 When initiated Unity might not have set the target framerate correctly and when disabling the scaler initially it would override the wrong framerate.
 
         /// <summary>
         /// Ensures settings are applied during startup.
@@ -13,6 +14,7 @@ namespace UnityEngine.AdaptivePerformance
         protected override void Awake()
         {
             base.Awake();
+            m_FirstTimeStart = 0;
             if (m_Settings == null)
                 return;
             ApplyDefaultSetting(m_Settings.scalerSettings.AdaptiveFramerate);
@@ -23,6 +25,11 @@ namespace UnityEngine.AdaptivePerformance
         /// </summary>
         protected override void OnDisabled()
         {
+            if (m_FirstTimeStart<2)
+            {
+                m_FirstTimeStart++;
+                return;
+            }
             Application.targetFrameRate = m_DefaultFPS;
         }
 
@@ -31,6 +38,9 @@ namespace UnityEngine.AdaptivePerformance
         /// </summary>
         protected override void OnEnabled()
         {
+            if (m_FirstTimeStart < 2)
+                return;
+
             m_DefaultFPS = Application.targetFrameRate;
             Application.targetFrameRate = (int)MaxBound;
         }